+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/radio-widget.ts
@@ -32,28 +32,6 @@ function getDefaultOptions() {
     };
 }
 
-const version3 = optional(object({major: constant(3), minor: number}));
-const parseRadioWidgetV3 = parseWidgetWithVersion(
-    version3,
-    constant("radio"),
-    object({
-        numCorrect: optional(number),
-        choices: array(
-            object({
-                content: defaulted(string, () => ""),
-                rationale: optional(string),
-                correct: optional(boolean),
-                isNoneOfTheAbove: optional(boolean),
-            }),
-        ),
-        hasNoneOfTheAbove: optional(boolean),
-        countChoices: optional(boolean),
-        randomize: optional(boolean),
-        multipleSelect: optional(boolean),
-        deselectEnabled: optional(boolean),
-    }),
-);
-
 const version2 = optional(object({major: constant(2), minor: number}));
 const parseRadioWidgetV2 = parseWidgetWithVersion(
     version2,
@@ -155,30 +133,6 @@ const parseRadioWidgetV0 = parseWidgetWithVersion(
 );
 
 // migrate functions
-export function migrateV2toV3(
-    widget: ParsedValue<typeof parseRadioWidgetV2>,
-): ParsedValue<typeof parseRadioWidgetV3> {
-    const {options} = widget;
-    return {
-        ...widget,
-        version: {major: 3, minor: 0},
-        options: {
-            numCorrect: options.numCorrect,
-            hasNoneOfTheAbove: options.hasNoneOfTheAbove,
-            countChoices: options.countChoices,
-            randomize: options.randomize,
-            multipleSelect: options.multipleSelect,
-            deselectEnabled: options.deselectEnabled,
-            choices: options.choices.map((choice) => ({
-                content: choice.content,
-                rationale: choice.clue,
-                correct: choice.correct,
-                isNoneOfTheAbove: choice.isNoneOfTheAbove,
-            })),
-        },
-    };
-}
-
 export function migrateV1ToV2(
     widget: ParsedValue<typeof parseRadioWidgetV1>,
 ): ParsedValue<typeof parseRadioWidgetV2> {
@@ -209,7 +163,6 @@ export function migrateV0ToV1(
     };
 }
 
-export const parseRadioWidget = versionedWidgetOptions(3, parseRadioWidgetV3)
-    .withMigrationFrom(2, parseRadioWidgetV2, migrateV2toV3)
+export const parseRadioWidget = versionedWidgetOptions(2, parseRadioWidgetV2)
     .withMigrationFrom(1, parseRadioWidgetV1, migrateV1ToV2)
     .withMigrationFrom(0, parseRadioWidgetV0, migrateV0ToV1).parser;