+++ packages/perseus-editor/src/widgets/radio/editor.tsx
@@ -1,17 +1,16 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import {
-    components,
-    BaseRadio,
-    Changeable,
-    iconTrash,
-} from "@khanacademy/perseus";
+import {BaseRadio, Changeable} from "@khanacademy/perseus";
 import {radioLogic, deriveNumCorrect} from "@khanacademy/perseus-core";
-import {Checkbox} from "@khanacademy/wonder-blocks-form";
+import Button from "@khanacademy/wonder-blocks-button";
+import {Strut} from "@khanacademy/wonder-blocks-layout";
+import Link from "@khanacademy/wonder-blocks-link";
+import {spacing, sizing} from "@khanacademy/wonder-blocks-tokens";
+import plusIcon from "@phosphor-icons/core/bold/plus-bold.svg";
 import * as React from "react";
 import _ from "underscore";
 
+import LabeledSwitch from "../../components/labeled-switch";
 import Editor from "../../editor";
-import {iconPlus} from "../../styles/icon-paths";
 
 import type {APIOptions} from "@khanacademy/perseus";
 import type {
@@ -20,14 +19,12 @@ import type {
     RadioDefaultWidgetOptions,
 } from "@khanacademy/perseus-core";
 
-const {InlineIcon} = components;
-
 type Contentful = {content?: string};
 type ChoiceEditorProps = {
     apiOptions: APIOptions;
     choice: PerseusRadioChoice;
     showDelete: boolean;
-    onRationaleChange: (newProps: Contentful) => void;
+    onClueChange: (newProps: Contentful) => void;
     onContentChange: (newProps: Contentful) => void;
     onDelete: () => void;
 };
@@ -61,18 +58,17 @@ class ChoiceEditor extends React.Component<ChoiceEditorProps> {
             />
         );
 
-        const rationaleEditor = (
+        const clueEditor = (
             <Editor
                 // eslint-disable-next-line react/no-string-refs
-                ref="rationale-editor"
+                ref="clue-editor"
                 apiOptions={this.props.apiOptions}
-                content={this.props.choice.rationale || ""}
+                content={this.props.choice.clue || ""}
                 widgetEnabled={false}
                 placeholder={`Why is this choice ${checkedClass}?`}
-                onChange={this.props.onRationaleChange}
+                onChange={this.props.onClueChange}
             />
         );
-
         const deleteLink = (
             <a
                 className="simple-button orange delete-choice"
@@ -84,14 +80,14 @@ class ChoiceEditor extends React.Component<ChoiceEditorProps> {
                 }}
                 title="Remove this choice"
             >
-                <InlineIcon {...iconTrash} />
+                Remove this choice
             </a>
         );
 
         return (
-            <div className="choice-rationale-editors">
+            <div className="choice-clue-editors">
                 <div className={`choice-editor ${checkedClass}`}>{editor}</div>
-                <div className="rationale-editor">{rationaleEditor}</div>
+                <div className="clue-editor">{clueEditor}</div>
                 {this.props.showDelete && deleteLink}
             </div>
         );
@@ -102,6 +98,7 @@ type RadioEditorProps = {
     apiOptions: APIOptions;
     countChoices: boolean;
     choices: PerseusRadioChoice[];
+    displayCount: number;
     randomize: boolean;
     hasNoneOfTheAbove: boolean;
     multipleSelect: boolean;
@@ -191,13 +188,13 @@ class RadioEditor extends React.Component<RadioEditorProps> {
         this.props.onChange({choices: choices});
     };
 
-    onRationaleChange(choiceIndex: number, newRationale: string): void {
+    onClueChange(choiceIndex: number, newClue: string): void {
         const choices = this.props.choices.slice();
         choices[choiceIndex] = _.extend({}, choices[choiceIndex], {
-            rationale: newRationale,
+            clue: newClue,
         });
-        if (newRationale === "") {
-            delete choices[choiceIndex].rationale;
+        if (newClue === "") {
+            delete choices[choiceIndex].clue;
         }
         this.props.onChange({choices: choices});
     }
@@ -244,6 +241,10 @@ class RadioEditor extends React.Component<RadioEditorProps> {
         );
     };
 
+    setDisplayCount: (arg1: number) => void = (num) => {
+        this.props.onChange({displayCount: num});
+    };
+
     focus: () => boolean = () => {
         // eslint-disable-next-line react/no-string-refs
         // @ts-expect-error - TS2339 - Property 'refs' does not exist on type 'ReactInstance'.
@@ -264,6 +265,7 @@ class RadioEditor extends React.Component<RadioEditorProps> {
             randomize,
             multipleSelect,
             countChoices,
+            displayCount,
             hasNoneOfTheAbove,
             deselectEnabled,
         } = this.props;
@@ -273,6 +275,7 @@ class RadioEditor extends React.Component<RadioEditorProps> {
             randomize,
             multipleSelect,
             countChoices,
+            displayCount,
             hasNoneOfTheAbove,
             deselectEnabled,
             numCorrect: deriveNumCorrect({
@@ -294,50 +297,42 @@ class RadioEditor extends React.Component<RadioEditorProps> {
         );
         return (
             <div>
+                <Link
+                    href="https://www.khanacademy.org/internal-courses/content-creation-best-practices/xe46daa512cd9c644:question-writing/xe46daa512cd9c644:multiple-choice/a/stems"
+                    target="_blank"
+                >
+                    Multiple choice best practices
+                </Link>
                 <div className="perseus-widget-row">
-                    <a
-                        href={
-                            // This is an editor component, not user-facing.
-                            "https://www.khanacademy.org/internal-courses/content-creation-best-practices/xe46daa512cd9c644:question-writing/xe46daa512cd9c644:multiple-choice/a/stems"
-                        }
-                        target="_blank"
-                        rel="noreferrer"
-                    >
-                        Multiple choice best practices
-                    </a>
-                    <br />
-                    <div className="perseus-widget-left-col">
-                        <Checkbox
-                            label="Multiple selections"
-                            checked={this.props.multipleSelect}
+                    <LabeledSwitch
+                        label="Randomize order"
+                        checked={this.props.randomize}
+                        onChange={(value) => {
+                            this.props.onChange({randomize: value});
+                        }}
+                        style={{marginBlockEnd: sizing.size_060}}
+                    />
+                    <LabeledSwitch
+                        label="Multiple selections"
+                        checked={this.props.multipleSelect}
+                        onChange={(value) => {
+                            this.onMultipleSelectChange({
+                                multipleSelect: value,
+                            });
+                        }}
+                        style={{marginBlockEnd: sizing.size_060}}
+                    />
+                    {this.props.multipleSelect && (
+                        <LabeledSwitch
+                            label="Specify number correct"
+                            checked={this.props.countChoices}
                             onChange={(value) => {
-                                this.onMultipleSelectChange({
-                                    multipleSelect: value,
+                                this.onCountChoicesChange({
+                                    countChoices: value,
                                 });
                             }}
+                            style={{marginBlockEnd: sizing.size_060}}
                         />
-                    </div>
-                    <div className="perseus-widget-right-col">
-                        <Checkbox
-                            label="Randomize order"
-                            checked={this.props.randomize}
-                            onChange={(value) => {
-                                this.props.onChange({randomize: value});
-                            }}
-                        />
-                    </div>
-                    {this.props.multipleSelect && (
-                        <div className="perseus-widget-left-col">
-                            <Checkbox
-                                label="Specify number correct"
-                                checked={this.props.countChoices}
-                                onChange={(value) => {
-                                    this.onCountChoicesChange({
-                                        countChoices: value,
-                                    });
-                                }}
-                            />
-                        </div>
                     )}
                 </div>
 
@@ -364,9 +359,9 @@ class RadioEditor extends React.Component<RadioEditorProps> {
                                             );
                                         }
                                     }}
-                                    onRationaleChange={(newProps) => {
+                                    onClueChange={(newProps) => {
                                         if (newProps.content != null) {
-                                            this.onRationaleChange(
+                                            this.onClueChange(
                                                 i,
                                                 newProps.content,
                                             );
@@ -384,25 +379,23 @@ class RadioEditor extends React.Component<RadioEditorProps> {
                 />
 
                 <div className="add-choice-container">
-                    <a
-                        className="simple-button orange"
-                        href="#"
-                        // eslint-disable-next-line react/jsx-no-bind
+                    <Button
+                        size="small"
+                        kind="tertiary"
+                        startIcon={plusIcon}
                         onClick={this.addChoice.bind(this, false)}
                     >
-                        <InlineIcon {...iconPlus} /> Add a choice{" "}
-                    </a>
-
-                    {!this.props.hasNoneOfTheAbove && (
-                        <a
-                            className="simple-button"
-                            href="#"
-                            // eslint-disable-next-line react/jsx-no-bind
-                            onClick={this.addChoice.bind(this, true)}
-                        >
-                            <InlineIcon {...iconPlus} /> None of the above{" "}
-                        </a>
-                    )}
+                        Add a choice
+                    </Button>
+                    <Strut size={spacing.large_24} />
+                    <Button
+                        size="small"
+                        kind="tertiary"
+                        startIcon={plusIcon}
+                        onClick={this.addChoice.bind(this, true)}
+                    >
+                        None of the above
+                    </Button>
                 </div>
             </div>
         );